+++ src/lib/notcurses.c
@@ -57,15 +57,15 @@ notcurses_stop_minimal(void* vnc){
   // be sure to write the restoration sequences *prior* to running rmcup, as
   // they apply to the screen (alternate or otherwise) we're actually using.
   const char* esc;
+  // ECMA-48 suggests that we can interrupt an escape code with a NUL
+  // byte. if we leave an active escape open, it can lock up the terminal.
+  // we only want to do it when in the middle of a rasterization, though. FIXME
+  if(nc->tcache.pixel_shutdown){
+    ret |= nc->tcache.pixel_shutdown(nc->ttyfp);
+  }
+  ret |= notcurses_mouse_disable(nc);
+  ret |= reset_term_attributes(&nc->tcache, nc->ttyfp);
   if(nc->ttyfd >= 0){
-    // ECMA-48 suggests that we can interrupt an escape code with a NUL
-    // byte. if we leave an active escape open, it can lock up the terminal.
-    // we only want to do it when in the middle of a rasterization, though. FIXME
-    if(nc->tcache.pixel_shutdown){
-      ret |= nc->tcache.pixel_shutdown(nc->ttyfd);
-    }
-    ret |= notcurses_mouse_disable(nc);
-    ret |= reset_term_attributes(&nc->tcache, nc->ttyfp);
     if((esc = get_escape(&nc->tcache, ESCAPE_RMCUP))){
       if(sprite_clear_all(&nc->tcache, nc->ttyfp)){
         ret = -1;